requirement.txt
nvidia-cuda-runtime-cu12
nvidia-cublas-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
fastapi
uvicorn[standard]
python-multipart
easyocr
pillow
numpy